+++ changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## 0.2.0
+- [CLI] Added help message.
+- [CLI] Added `--full` function to display the full list of versions when only an edition or phase is given, which is now hidden by default.
 - [Data] Added Java Beta versions.
 
 ## 0.1.3